requirements.txt
# fastapi
# uvicorn
# google-generativeai
# python-dotenv
# google-search-results
# langchain
# langchain_google_genai
# langchain_openai

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.10
langchain-openai==0.1.22
openai==1.40.1
google-search-results==2.4.2  # SerpAPI SDK